chore: set line-height of black pre-commit settings to 120 and run changes

+++ argilla-server/tests/unit/api/v1/test_fields.py
@@ -92,9 +92,7 @@ async def test_update_field_with_invalid_title(
 ):
     field = await TextFieldFactory.create(title="title")
 
-    response = await async_client.patch(
-        f"/api/v1/fields/{field.id}", headers=owner_auth_header, json={"title": title}
-    )
+    response = await async_client.patch(f"/api/v1/fields/{field.id}", headers=owner_auth_header, json={"title": title})
 
     assert response.status_code == 422
 
@@ -112,22 +110,16 @@ async def test_update_field_with_invalid_title(
     ],
 )
 @pytest.mark.asyncio
-async def test_update_field_with_invalid_settings(
-    async_client: "AsyncClient", owner_auth_header: dict, payload: dict
-):
+async def test_update_field_with_invalid_settings(async_client: "AsyncClient", owner_auth_header: dict, payload: dict):
     field = await TextFieldFactory.create()
 
-    response = await async_client.patch(
-        f"/api/v1/fields/{field.id}", headers=owner_auth_header, json=payload
-    )
+    response = await async_client.patch(f"/api/v1/fields/{field.id}", headers=owner_auth_header, json=payload)
 
     assert response.status_code == 422
 
 
 @pytest.mark.asyncio
-async def test_update_field_with_invalid_payload(
-    async_client: "AsyncClient", owner_auth_header: dict
-):
+async def test_update_field_with_invalid_payload(async_client: "AsyncClient", owner_auth_header: dict):
     field = await TextFieldFactory.create()
 
     response = await async_client.patch(
@@ -143,9 +135,7 @@ async def test_update_field_with_invalid_payload(
 
 
 @pytest.mark.asyncio
-async def test_update_field_non_existent(
-    async_client: "AsyncClient", owner_auth_header: dict
-):
+async def test_update_field_non_existent(async_client: "AsyncClient", owner_auth_header: dict):
     field_id = uuid4()
 
     response = await async_client.patch(
@@ -177,9 +167,7 @@ async def test_update_field_as_admin_from_different_workspace(
 @pytest.mark.asyncio
 async def test_update_field_as_annotator(async_client: "AsyncClient"):
     field = await TextFieldFactory.create()
-    user = await UserFactory.create(
-        role=UserRole.annotator, workspaces=[field.dataset.workspace]
-    )
+    user = await UserFactory.create(role=UserRole.annotator, workspaces=[field.dataset.workspace])
 
     response = await async_client.patch(
         f"/api/v1/fields/{field.id}",
@@ -203,14 +191,10 @@ async def test_update_field_without_authentication(async_client: "AsyncClient"):
 
 
 @pytest.mark.asyncio
-async def test_delete_field(
-    async_client: "AsyncClient", db: "AsyncSession", owner_auth_header: dict
-):
+async def test_delete_field(async_client: "AsyncClient", db: "AsyncSession", owner_auth_header: dict):
     field = await TextFieldFactory.create(name="name", title="title")
 
-    response = await async_client.delete(
-        f"/api/v1/fields/{field.id}", headers=owner_auth_header
-    )
+    response = await async_client.delete(f"/api/v1/fields/{field.id}", headers=owner_auth_header)
 
     assert response.status_code == 200
     assert (await db.execute(select(func.count(Field.id)))).scalar() == 0
@@ -229,9 +213,7 @@ async def test_delete_field(
 
 
 @pytest.mark.asyncio
-async def test_delete_field_without_authentication(
-    async_client: "AsyncClient", db: "AsyncSession"
-):
+async def test_delete_field_without_authentication(async_client: "AsyncClient", db: "AsyncSession"):
     field = await TextFieldFactory.create()
 
     response = await async_client.delete(f"/api/v1/fields/{field.id}")
@@ -241,15 +223,11 @@ async def test_delete_field_without_authentication(
 
 
 @pytest.mark.asyncio
-async def test_delete_field_as_annotator(
-    async_client: "AsyncClient", db: "AsyncSession"
-):
+async def test_delete_field_as_annotator(async_client: "AsyncClient", db: "AsyncSession"):
     annotator = await AnnotatorFactory.create()
     field = await TextFieldFactory.create()
 
-    response = await async_client.delete(
-        f"/api/v1/fields/{field.id}", headers={API_KEY_HEADER_NAME: annotator.api_key}
-    )
+    response = await async_client.delete(f"/api/v1/fields/{field.id}", headers={API_KEY_HEADER_NAME: annotator.api_key})
 
     assert response.status_code == 403
     assert (await db.execute(select(func.count(Field.id)))).scalar() == 1
@@ -262,14 +240,10 @@ async def test_delete_field_belonging_to_published_dataset(
     dataset = await DatasetFactory.create(status=DatasetStatus.ready)
     field = await TextFieldFactory.create(dataset=dataset)
 
-    response = await async_client.delete(
-        f"/api/v1/fields/{field.id}", headers=owner_auth_header
-    )
+    response = await async_client.delete(f"/api/v1/fields/{field.id}", headers=owner_auth_header)
 
     assert response.status_code == 422
-    assert response.json() == {
-        "detail": "Fields cannot be deleted for a published dataset"
-    }
+    assert response.json() == {"detail": "Fields cannot be deleted for a published dataset"}
 
     assert (await db.execute(select(func.count(Field.id)))).scalar() == 1
 
@@ -282,9 +256,7 @@ async def test_delete_field_with_nonexistent_field_id(
 
     await TextFieldFactory.create()
 
-    response = await async_client.delete(
-        f"/api/v1/fields/{field_id}", headers=owner_auth_header
-    )
+    response = await async_client.delete(f"/api/v1/fields/{field_id}", headers=owner_auth_header)
 
     assert response.status_code == 404
     assert response.json() == {"detail": f"Field with id `{field_id}` not found"}
